pull out closes feature calc, remove rep station file
